+++ jetwhale-agent-runtime/src/commonMain/kotlin/com/kitakkun/jetwhale/agent/runtime/DefaultJetWhaleMessagingService.kt
@@ -68,11 +68,6 @@ internal class DefaultJetWhaleMessagingService(
 
                 is JetWhaleDebuggerEvent.PluginDeactivated -> pluginService.deactivatePlugins(event.pluginId)
 
-                is JetWhaleDebuggerEvent.PluginEvent -> {
-                    val plugin = pluginService.getPluginById(event.pluginId) ?: return@collect
-                    plugin.onRawDebuggerEvent(event.payload)
-                }
-
                 is JetWhaleDebuggerEvent.MethodRequest -> {
                     val plugin = pluginService.getPluginById(event.pluginId) ?: return@collect
                     val methodResult = plugin.onRawMethod(event.payload)
