+++ jetwhale-agent-runtime/src/commonMain/kotlin/com/kitakkun/jetwhale/agent/runtime/JetWhaleServiceDsl.kt
@@ -1,7 +1,5 @@
 package com.kitakkun.jetwhale.agent.runtime
 
-import com.kitakkun.jetwhale.agent.sdk.JetWhaleAgentPlugin
-import com.kitakkun.jetwhale.agent.sdk.JetWhaleMethodAgentPlugin
 import com.kitakkun.jetwhale.annotations.InternalJetWhaleApi
 import com.kitakkun.jetwhale.protocol.serialization.JetWhaleJson
 import io.ktor.client.HttpClient
@@ -62,8 +60,7 @@ public interface JetWhaleLoggingConfigurationScope {
 
 @JetWhaleDsl
 public interface JetWhalePluginConfigurationScope {
-    public fun register(plugin: JetWhaleAgentPlugin<*, *>)
-    public fun register(plugin: JetWhaleMethodAgentPlugin<*, *, *>)
+    public fun register(plugin: AgentPlugin)
 }
 
 private class JetWhaleConfiguration : JetWhaleConfigurationScope {
@@ -95,15 +92,10 @@ private class JetWhaleLoggingConfiguration : JetWhaleLoggingConfigurationScope {
     override var ktorLogLevel: KtorLogLevel = KtorLogLevel.NONE
 }
 
-@OptIn(InternalJetWhaleApi::class)
 private class JetWhalePluginConfiguration : JetWhalePluginConfigurationScope {
     val plugins: MutableList<AgentPlugin> = mutableListOf()
 
-    override fun register(plugin: JetWhaleAgentPlugin<*, *>) {
-        plugins.add(plugin)
-    }
-
-    override fun register(plugin: JetWhaleMethodAgentPlugin<*, *, *>) {
+    override fun register(plugin: AgentPlugin) {
         plugins.add(plugin)
     }
 }
