+++ jetwhale-agent-sdk/src/commonMain/kotlin/com/kitakkun/jetwhale/agent/sdk/JetWhaleAgentPlugin.kt
@@ -4,55 +4,46 @@ import com.kitakkun.jetwhale.annotations.InternalJetWhaleApi
 import com.kitakkun.jetwhale.protocol.agent.JetWhaleAgentPluginProtocol
 
 /**
- * Standard bidirectional event-based plugin which runs on the debug-target app.
- *
- * Supports fire-and-forget communication in both directions:
- * - [enqueueEvent]: send a [DebuggeeEvent] from the agent to the debugger
- * - [onReceiveDebuggerEvent]: receive a [DebuggerEvent] from the debugger
- *
- * For request-response communication via Method/MethodResult, use [JetWhaleMethodAgentPlugin] instead.
+ * Plugin client which runs on the debug-target app.
  */
 @OptIn(InternalJetWhaleApi::class)
-public abstract class JetWhaleAgentPlugin<DebuggeeEvent, DebuggerEvent> : JetWhaleRawAgentPlugin() {
+public abstract class JetWhaleAgentPlugin<Event, Method, MethodResult> : JetWhaleRawAgentPlugin() {
     /**
      * The protocol used for encoding and decoding messages.
      */
-    protected abstract val protocol: JetWhaleAgentPluginProtocol<DebuggeeEvent, DebuggerEvent>
+    protected abstract val protocol: JetWhaleAgentPluginProtocol<Event, Method, MethodResult>
 
     /**
-     * Handles a typed event received from the debugger.
-     * @param event The decoded debugger event.
+     * Handles a typed method message received from the debugger.
+     * @param method The decoded method message.
+     * @return The result of the method, or null if no response is needed.
      */
-    public abstract suspend fun onReceiveDebuggerEvent(event: DebuggerEvent)
+    public abstract suspend fun onReceiveMethod(method: Method): MethodResult?
 
     /**
      * Called when an event is enqueued.
      * Mainly for logging or debugging purposes.
      * @param event The event message that was enqueued.
      */
-    public open fun onEnqueueEvent(event: DebuggeeEvent) {}
+    public open fun onEnqueueEvent(event: Event) {}
 
     /**
-     * Enqueues a typed event to be sent to the debugger.
-     * @param event The event to be sent.
+     * Enqueues a typed event message to be sent to the debugger.
+     * @param event The event message to be sent.
      */
-    public fun enqueueEvent(event: DebuggeeEvent) {
-        val rawPayload = protocol.encodeDebuggeeEvent(event)
+    public fun enqueueEvent(event: Event) {
+        val rawPayload = protocol.encodeEvent(event)
         enqueueRawEvent(rawPayload)
         onEnqueueEvent(event)
     }
 
     /**
-     * Handles a raw fire-and-forget event from the debugger.
-     * Decodes and dispatches to [onReceiveDebuggerEvent].
+     * Handles a raw method message received from the debugger.
+     * Decodes the message, processes it, and encodes the result.
      */
-    final override suspend fun onRawDebuggerEvent(message: String) {
-        val event = protocol.decodeDebuggerEvent(message)
-        onReceiveDebuggerEvent(event)
+    final override suspend fun onRawMethod(message: String): String? {
+        val method = protocol.decodeMethod(message)
+        val result = onReceiveMethod(method) ?: return null
+        return protocol.encodeMethodResult(result)
     }
-
-    /**
-     * Not used by this plugin type. Method-based request-response is handled by [JetWhaleMethodAgentPlugin].
-     */
-    final override suspend fun onRawMethod(message: String): String? = null
 }
