+++ jetwhale-host-sdk/src/main/kotlin/com/kitakkun/jetwhale/host/sdk/JetWhaleHostPlugin.kt
@@ -5,25 +5,19 @@ import com.kitakkun.jetwhale.protocol.host.JetWhaleHostPluginProtocol
 import kotlinx.coroutines.CoroutineScope
 
 /**
- * Standard bidirectional event-based plugin interface for JetWhale Host.
- *
- * Supports fire-and-forget communication in both directions:
- * - [onEvent]: receive a [DebuggeeEvent] from the debuggee
- * - [JetWhaleDebugOperationContext.send]: send a [DebuggerEvent] to the debuggee
- *
- * For request-response communication via Method/MethodResult, use [JetWhaleMethodHostPlugin] instead.
+ * Plugin interface for JetWhale.
  */
-public abstract class JetWhaleHostPlugin<DebuggeeEvent, DebuggerEvent> : JetWhaleRawHostPlugin() {
+public abstract class JetWhaleHostPlugin<Event, Method, MethodResult> : JetWhaleRawHostPlugin() {
     /**
-     * The protocol used for encoding and decoding events
+     * The protocol used for encoding and decoding events, methods, and method results
      */
-    protected abstract val protocol: JetWhaleHostPluginProtocol<DebuggeeEvent, DebuggerEvent>
+    protected abstract val protocol: JetWhaleHostPluginProtocol<Event, Method, MethodResult>
 
     /**
      * Called when an event is received from debuggee
      * @param event The received event
      */
-    public abstract fun onEvent(event: DebuggeeEvent)
+    public abstract fun onEvent(event: Event)
 
     /**
      * Composable function that represents the UI of the plugin
@@ -31,17 +25,17 @@ public abstract class JetWhaleHostPlugin<DebuggeeEvent, DebuggerEvent> : JetWhal
      * Composition will be kept as long as the plugin instance is alive.
      * Even if the tab is switched, the composition will be kept.
      *
-     * @param context The context to send debugger events
+     * @param context The context to perform debug operations
      */
     @Composable
-    public abstract fun Content(context: JetWhaleDebugOperationContext<DebuggerEvent>)
+    public abstract fun Content(context: JetWhaleDebugOperationContext<Method, MethodResult>)
 
     /**
      * Handles a raw event message received from the debuggee.
      * Decodes the message and processes it.
      */
     final override fun onRawEvent(event: String) {
-        val decodedEvent = protocol.decodeDebuggeeEvent(event)
+        val decodedEvent = protocol.decodeEvent(event)
         onEvent(decodedEvent)
     }
 
@@ -52,10 +46,13 @@ public abstract class JetWhaleHostPlugin<DebuggeeEvent, DebuggerEvent> : JetWhal
      */
     @Composable
     final override fun ContentRaw(context: JetWhaleRawDebugOperationContext) {
-        val typedContext = object : JetWhaleDebugOperationContext<DebuggerEvent> {
+        val typedContext = object : JetWhaleDebugOperationContext<Method, MethodResult> {
             override val coroutineScope: CoroutineScope = context.coroutineScope
-            override suspend fun send(event: DebuggerEvent) {
-                context.send(protocol.encodeDebuggerEvent(event))
+            override suspend fun <MR : MethodResult> dispatch(method: Method): MR? {
+                val encodedMethod = protocol.encodeMethod(method)
+                val rawResult = context.dispatch(encodedMethod)
+                @Suppress("UNCHECKED_CAST")
+                return rawResult?.let { protocol.decodeMethodResult(it) } as? MR
             }
         }
         Content(context = typedContext)
