--- jetwhale-host-sdk/src/main/kotlin/com/kitakkun/jetwhale/host/sdk/JetWhaleMethodHostPlugin.kt
@@ -1,63 +1,0 @@
-package com.kitakkun.jetwhale.host.sdk
-
-import androidx.compose.runtime.Composable
-import com.kitakkun.jetwhale.protocol.host.JetWhaleMethodHostPluginProtocol
-import kotlinx.coroutines.CoroutineScope
-
-/**
- * Method-based plugin interface for JetWhale Host.
- * Supports request-response communication via Method and MethodResult types.
- *
- * For simple bidirectional fire-and-forget event communication, use [JetWhaleHostPlugin] instead.
- */
-public abstract class JetWhaleMethodHostPlugin<Event, Method, MethodResult> : JetWhaleRawHostPlugin() {
-    /**
-     * The protocol used for encoding and decoding events, methods, and method results
-     */
-    protected abstract val protocol: JetWhaleMethodHostPluginProtocol<Event, Method, MethodResult>
-
-    /**
-     * Called when an event is received from debuggee
-     * @param event The received event
-     */
-    public abstract fun onEvent(event: Event)
-
-    /**
-     * Composable function that represents the UI of the plugin
-     *
-     * Composition will be kept as long as the plugin instance is alive.
-     * Even if the tab is switched, the composition will be kept.
-     *
-     * @param context The context to perform debug operations
-     */
-    @Composable
-    public abstract fun Content(context: JetWhaleMethodDebugOperationContext<Method, MethodResult>)
-
-    /**
-     * Handles a raw event message received from the debuggee.
-     * Decodes the message and processes it.
-     */
-    final override fun onRawEvent(event: String) {
-        val decodedEvent = protocol.decodeEvent(event)
-        onEvent(decodedEvent)
-    }
-
-    /**
-     * Composable function that adapts raw string-based context to typed context
-     * and calls the typed Content function.
-     * @param context The raw debug operation context
-     */
-    @Composable
-    final override fun ContentRaw(context: JetWhaleRawDebugOperationContext) {
-        val typedContext = object : JetWhaleMethodDebugOperationContext<Method, MethodResult> {
-            override val coroutineScope: CoroutineScope = context.coroutineScope
-            override suspend fun <MR : MethodResult> dispatch(method: Method): MR? {
-                val encodedMethod = protocol.encodeMethod(method)
-                val rawResult = context.dispatch(encodedMethod)
-                @Suppress("UNCHECKED_CAST")
-                return rawResult?.let { protocol.decodeMethodResult(it) } as? MR
-            }
-        }
-        Content(context = typedContext)
-    }
-}
