+++ jetwhale-host/core/data/src/main/kotlin/com/kitakkun/jetwhale/host/data/plugin/DefaultPluginComposeSceneService.kt
@@ -57,11 +57,6 @@ class DefaultPluginComposeSceneService(
                         sessionId = sessionId,
                         payload = method,
                     )
-                    override suspend fun send(encodedEvent: String) = debugWebSocketServer.sendEvent(
-                        pluginId = pluginId,
-                        sessionId = sessionId,
-                        payload = encodedEvent,
-                    )
                 }
 
                 val windowUpdatableContext = DynamicWindowInfoPlatformContext()
