+++ jetwhale-plugins/example/agent/src/commonMain/kotlin/com/kitakkun/jetwhale/plugins/example/agent/ExampleAgentPlugin.kt
@@ -1,11 +1,11 @@
 package com.kitakkun.jetwhale.plugins.example.agent
 
-import com.kitakkun.jetwhale.agent.sdk.JetWhaleMethodAgentPlugin
+import com.kitakkun.jetwhale.agent.sdk.JetWhaleAgentPlugin
 import com.kitakkun.jetwhale.plugins.example.protocol.ExampleEvent
 import com.kitakkun.jetwhale.plugins.example.protocol.ExampleMethod
 import com.kitakkun.jetwhale.plugins.example.protocol.ExampleMethodResult
-import com.kitakkun.jetwhale.protocol.agent.JetWhaleMethodAgentPluginProtocol
-import com.kitakkun.jetwhale.protocol.agent.kotlinxSerializationJetWhaleMethodAgentPluginProtocol
+import com.kitakkun.jetwhale.protocol.agent.JetWhaleAgentPluginProtocol
+import com.kitakkun.jetwhale.protocol.agent.kotlinxSerializationJetWhaleAgentPluginProtocol
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.update
@@ -14,10 +14,10 @@ import kotlinx.coroutines.flow.update
  * An example JetWhale agent plugin that handles ExampleMethod and ExampleEvent.
  * It logs received methods and events to a StateFlow of event logs.
  */
-class ExampleAgentPlugin : JetWhaleMethodAgentPlugin<ExampleEvent, ExampleMethod, ExampleMethodResult>() {
+class ExampleAgentPlugin : JetWhaleAgentPlugin<ExampleEvent, ExampleMethod, ExampleMethodResult>() {
     override val pluginId: String get() = "com.kitakkun.jetwhale.example"
     override val pluginVersion: String get() = "1.0.0"
-    override val protocol: JetWhaleMethodAgentPluginProtocol<ExampleEvent, ExampleMethod, ExampleMethodResult> = kotlinxSerializationJetWhaleMethodAgentPluginProtocol()
+    override val protocol: JetWhaleAgentPluginProtocol<ExampleEvent, ExampleMethod, ExampleMethodResult> = kotlinxSerializationJetWhaleAgentPluginProtocol()
 
     val mutableEventLogsFlow: MutableStateFlow<List<String>> = MutableStateFlow(emptyList())
     val eventLogsFlow: StateFlow<List<String>> = mutableEventLogsFlow
