+++ jetwhale-plugins/example/host/src/main/kotlin/com/kitakkun/jetwhale/plugins/example/host/ExamplePluginUI.kt
@@ -15,7 +15,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.snapshots.SnapshotStateList
 import androidx.compose.ui.Modifier
-import com.kitakkun.jetwhale.host.sdk.JetWhaleMethodDebugOperationContext
+import com.kitakkun.jetwhale.host.sdk.JetWhaleDebugOperationContext
 import com.kitakkun.jetwhale.plugins.example.protocol.ExampleMethod
 import com.kitakkun.jetwhale.plugins.example.protocol.ExampleMethodResult
 import kotlinx.coroutines.launch
@@ -23,7 +23,7 @@ import kotlinx.coroutines.launch
 @Composable
 fun ExamplePluginContent(
     eventLogs: SnapshotStateList<String>,
-    context: JetWhaleMethodDebugOperationContext<ExampleMethod, ExampleMethodResult>,
+    context: JetWhaleDebugOperationContext<ExampleMethod, ExampleMethodResult>,
 ) {
     ExamplePluginView(
         eventLogs = eventLogs,
