--- jetwhale-protocol/agent/src/commonMain/kotlin/com/kitakkun/jetwhale/protocol/agent/JetWhaleMethodAgentPluginProtocol.kt
@@ -1,37 +1,0 @@
-package com.kitakkun.jetwhale.protocol.agent
-
-import com.kitakkun.jetwhale.annotations.InternalJetWhaleApi
-import com.kitakkun.jetwhale.protocol.serialization.JetWhaleJson
-import kotlinx.serialization.KSerializer
-import kotlinx.serialization.json.Json
-import kotlinx.serialization.serializer
-
-public interface JetWhaleMethodAgentPluginProtocol<Event, Method, MethodResult> {
-    public fun decodeMethod(value: String): Method
-    public fun encodeMethodResult(value: MethodResult): String
-    public fun encodeEvent(value: Event): String
-}
-
-@PublishedApi
-internal class KotlinxSerializationJetWhaleMethodAgentPluginProtocol<Event, Method, MethodResult>(
-    private val json: Json,
-    private val eventSerializer: KSerializer<Event>,
-    private val methodSerializer: KSerializer<Method>,
-    private val methodResultSerializer: KSerializer<MethodResult>,
-) : JetWhaleMethodAgentPluginProtocol<Event, Method, MethodResult> {
-    override fun decodeMethod(value: String): Method = json.decodeFromString(methodSerializer, value)
-
-    override fun encodeMethodResult(value: MethodResult): String = json.encodeToString(methodResultSerializer, value)
-
-    override fun encodeEvent(value: Event): String = json.encodeToString(eventSerializer, value)
-}
-
-@OptIn(InternalJetWhaleApi::class)
-public inline fun <reified Event : Any, reified Method : Any, reified MethodResult : Any> kotlinxSerializationJetWhaleMethodAgentPluginProtocol(
-    json: Json = JetWhaleJson,
-): JetWhaleMethodAgentPluginProtocol<Event, Method, MethodResult> = KotlinxSerializationJetWhaleMethodAgentPluginProtocol(
-    json = json,
-    eventSerializer = serializer(),
-    methodSerializer = serializer(),
-    methodResultSerializer = serializer(),
-)
