+++ jetwhale-protocol/host/src/main/kotlin/com/kitakkun/jetwhale/protocol/host/JetWhaleHostPluginProtocol.kt
@@ -6,27 +6,32 @@ import kotlinx.serialization.KSerializer
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.serializer
 
-public interface JetWhaleHostPluginProtocol<DebuggeeEvent, DebuggerEvent> {
-    public fun decodeDebuggeeEvent(value: String): DebuggeeEvent
-    public fun encodeDebuggerEvent(value: DebuggerEvent): String
+public interface JetWhaleHostPluginProtocol<Event, Method, MethodResult> {
+    public fun encodeMethod(value: Method): String
+    public fun decodeMethodResult(value: String): MethodResult
+    public fun decodeEvent(value: String): Event
 }
 
 @PublishedApi
-internal class KotlinxSerializationJetWhaleHostPluginProtocol<DebuggeeEvent, DebuggerEvent>(
+internal class KotlinxSerializationJetWhaleHostPluginProtocol<Event, Method, MethodResult>(
     private val json: Json,
-    private val debuggeeEventSerializer: KSerializer<DebuggeeEvent>,
-    private val debuggerEventSerializer: KSerializer<DebuggerEvent>,
-) : JetWhaleHostPluginProtocol<DebuggeeEvent, DebuggerEvent> {
-    override fun decodeDebuggeeEvent(value: String): DebuggeeEvent = json.decodeFromString(debuggeeEventSerializer, value)
+    private val eventSerializer: KSerializer<Event>,
+    private val methodSerializer: KSerializer<Method>,
+    private val methodResultSerializer: KSerializer<MethodResult>,
+) : JetWhaleHostPluginProtocol<Event, Method, MethodResult> {
+    override fun encodeMethod(value: Method): String = json.encodeToString(methodSerializer, value)
 
-    override fun encodeDebuggerEvent(value: DebuggerEvent): String = json.encodeToString(debuggerEventSerializer, value)
+    override fun decodeMethodResult(value: String): MethodResult = json.decodeFromString(methodResultSerializer, value)
+
+    override fun decodeEvent(value: String): Event = json.decodeFromString(eventSerializer, value)
 }
 
 @OptIn(InternalJetWhaleApi::class)
-public inline fun <reified DebuggeeEvent : Any, reified DebuggerEvent : Any> kotlinxSerializationJetWhaleHostPluginProtocol(
+public inline fun <reified Event : Any, reified Method : Any, reified MethodResult : Any> kotlinxSerializationJetWhaleHostPluginProtocol(
     json: Json = JetWhaleJson,
-): JetWhaleHostPluginProtocol<DebuggeeEvent, DebuggerEvent> = KotlinxSerializationJetWhaleHostPluginProtocol(
+): JetWhaleHostPluginProtocol<Event, Method, MethodResult> = KotlinxSerializationJetWhaleHostPluginProtocol(
     json = json,
-    debuggeeEventSerializer = serializer(),
-    debuggerEventSerializer = serializer(),
+    eventSerializer = serializer(),
+    methodSerializer = serializer(),
+    methodResultSerializer = serializer(),
 )
